Move matches under transactions group
CLI: matches * → transactions matches *. The interactive 'review'
subcommand will be promoted to 'transactions review' in a follow-up
task that unifies it with the categorize review queue.
