+++ src/moneybin/cli/commands/transactions/__init__.py
@@ -1,12 +1,16 @@
 """Transactions top-level command group.
 
-Owns transaction operations (list, show, edit, bulk-edit) and
-per-transaction workflows (categorize, review matches) per cli-restructure.md v2.
+Owns transaction entity operations and workflows on transactions
+(matches, categorize, review) per cli-restructure.md v2.
 """
 
 import typer
+
+from . import matches
 
 app = typer.Typer(
     help="Transactions and workflows on them (matches, categorize, review)",
     no_args_is_help=True,
 )
+
+app.add_typer(matches.app, name="matches")
