+++ tests/moneybin/test_cli/test_help_no_wizard.py
@@ -30,7 +30,7 @@ _GROUPS = [
     ["import"],
     ["sync"],
     ["categorize"],
-    ["matches"],
+    ["transactions", "matches"],
     ["transform"],
     ["synthetic"],
     ["track"],
